requirements.txt
streamlit==1.28.2
google-generativeai
python-dotenv
pdf2image

